fix: moved Reasons class to _reference.py
