+++ src/library_analyzer/processing/api/purity_analysis/model/_call_graph.py
@@ -1,14 +1,17 @@
 from __future__ import annotations
 
 from dataclasses import dataclass, field
+from typing import TYPE_CHECKING
 
 from library_analyzer.processing.api.purity_analysis.model._module_data import (
     ClassScope,
     FunctionScope,
     NodeID,
-    Reasons,
     Reference,
 )
+
+if TYPE_CHECKING:
+    from library_analyzer.processing.api.purity_analysis.model._reference import Reasons
 
 
 @dataclass
